fixing the change the eap value and quantity

+++ sgqcmm/orcs/calc/utils/materiais_orcamento.py
@@ -380,7 +380,7 @@ class Selante():
         ).descricao
     
     def calcular_quantidade(self, largura, comprimento, repeticoes, estrutura_retratil):
-        dimensoes = 2 * (largura + comprimento) * repeticoes
+        dimensoes = Decimal(2 * (largura + comprimento) * repeticoes)
         if estrutura_retratil:
             self.quantidade = arrend_cima(Decimal(1.5) * dimensoes / 40, 0) * 2
         else:
